JWT authentication update

+++ backend/controller/prayerController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require("express-async-handler");
 const Prayer = require("../models/prayersModel");
+const User = require("../models/userModel");
 
 const getPrayer = asyncHandler(async (req, res) => {
-  const prayers = await Prayer.find();
+  const prayers = await Prayer.find({ user: req.user.id });
 
   res.json(prayers);
 });
@@ -14,6 +15,7 @@ const postPrayer = asyncHandler(async (req, res) => {
   }
   const prayers = await Prayer.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.json(prayers);
@@ -25,6 +27,19 @@ const putPrayer = asyncHandler(async (req, res) => {
   if (!prayers) {
     res.status(400);
     throw new Error("Prayer not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const updatedPrayer = await Prayer.findByIdAndUpdate(
@@ -46,7 +61,19 @@ const deletePrayer = asyncHandler(async (req, res) => {
     throw new Error("Prayer not found");
   }
 
-  const deletedPrayer = await Prayer.findByIdAndRemove(req.params.id);
+  // Check for user
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
+  await Prayer.findByIdAndRemove(req.params.id);
 
   res.json({ id: req.params.id });
 });
